fix: solved dispatch issue perf: separation of client component and server components doc: updated `README.md`
- removing `package-lock.json`
